requirements.txt
# Core
fastapi==0.103.0
pydantic==1.10.13
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
aiohttp==3.11.18
requests==2.31.0
httpx==0.28.0  # ✅ Compatible with supabase
supabase==2.15.3
jinja2==3.1.3
starlette==0.27.0

# Web scraping & parsing
beautifulsoup4>=4.11.0
tavily-python>=0.1.0
nest-asyncio>=1.5.0
pandas>=1.5.0

# Auth & Security
authlib==1.3.0
itsdangerous>=2.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
oauthlib>=3.2.0

# Task queue & scheduling
dramatiq==1.15.0
dramatiq[redis]==1.15.0
apscheduler>=3.10.0
schedule==1.2.1

# OpenAI
openai>=0.27.0



# Redis
redis==5.0.1

# Email & PDF
reportlab>=3.6.0
premailer>=3.10.0

# Utilities
tenacity>=8.2.0
python-dateutil==2.8.2
pytz>=2022.0
loguru>=0.6.0

# Server
gunicorn==21.2.0
